Allow filtering of arbitrary path params in the endpoint: tag
See README changes with this commit for usage notes.  This solves the problem of
path params that are not easily matchable with regex creating many tags for a
metric in datadog when you really want that endpoint to be one tag on a metric.

A good example is if you are using variable length human-readable slugs in your
request paths.